requirements.txt
funcy=1.15
matplotlib=3.3.2
tensorflow-gpu=2.3.1
